refactor: update
update change int64 to int32

+++ app/user/service/internal/biz/do.go
@@ -13,7 +13,7 @@ type Profile struct {
 
 type ProfileUpdate struct {
 	Profile
-	Status int64
+	Status int32
 }
 
 type Credentials struct {
